+++ keymanager/workload_service/server_test.go
@@ -173,7 +173,7 @@ func TestHandleGenerateKeySuccess(t *testing.T) {
 	if resp.KeyProtectionMechanism != keymanager.KeyProtectionMechanism_KEY_PROTECTION_VM_EMULATED.String() {
 		t.Fatalf("expected %s, got %s", keymanager.KeyProtectionMechanism_KEY_PROTECTION_VM_EMULATED.String(), resp.KeyProtectionMechanism)
 	}
-	if resp.ExpirationTime <= time.Now().Unix() {
+	if resp.ExpirationTime <= int32(time.Now().Unix()) {
 		t.Fatalf("expected expiration time in the future, got %d", resp.ExpirationTime)
 	}
 
@@ -529,7 +529,7 @@ func TestHandleEnumerateKeysWithKeys(t *testing.T) {
 		t.Fatalf("expected key protection mechanism %s, got %s", keymanager.KeyProtectionMechanism_KEY_PROTECTION_VM_EMULATED.String(), info1.KeyProtectionMechanism)
 	}
 	// Approximate check for expiration time
-	if info1.ExpirationTime <= time.Now().Unix() {
+	if info1.ExpirationTime <= int32(time.Now().Unix()) {
 		t.Fatalf("expected expiration time in the future, got %d", info1.ExpirationTime)
 	}
 
@@ -539,7 +539,7 @@ func TestHandleEnumerateKeysWithKeys(t *testing.T) {
 		t.Fatalf("expected kem2 %s in response", kem2)
 	}
 	// Approximate check for expiration time
-	if info2.ExpirationTime <= time.Now().Unix() {
+	if info2.ExpirationTime <= int32(time.Now().Unix()) {
 		t.Fatalf("expected expiration time in the future, got %d", info2.ExpirationTime)
 	}
 }
